+++ src/js/edu-benefits/utils/helpers.jsx
@@ -157,9 +157,3 @@ export function showYesNo(field) {
 export function enumToNames(enumValues, names) {
   return enumValues.map(item => names[item]);
 }
-
-export const hoursTypeLabels = {
-  semester: 'Semester',
-  quarter: 'Quarter',
-  clock: 'Clock'
-};
