+++ test/edu-benefits/5490/config/educationHistory.unit.spec.jsx
@@ -86,31 +86,4 @@ describe('Edu 5490 educationHistory', () => {
 
     expect(onSubmit.called).to.be.true;
   });
-
-  it('should add another', () => {
-    const onSubmit = sinon.spy();
-    const form = ReactTestUtils.renderIntoDocument(
-      <DefinitionTester
-          schema={schema}
-          onSubmit={onSubmit}
-          data={{}}
-          definitions={formConfig.defaultDefinitions}
-          uiSchema={uiSchema}/>
-    );
-    const formDOM = findDOMNode(form);
-    ReactTestUtils.Simulate.change(formDOM.querySelector('select'), {
-      target: {
-        value: 'graduated'
-      }
-    });
-    ReactTestUtils.Simulate.change(formDOM.querySelector('#root_postHighSchoolTrainings_0_name'), {
-      target: {
-        value: 'A college name'
-      }
-    });
-    ReactTestUtils.Simulate.click(formDOM.querySelector('.va-growable-add-btn'));
-
-    expect(formDOM.querySelector('.va-growable-background').textContent)
-      .to.contain('A college name');
-  });
 });
