+++ test/edu-benefits/definitions/postHighSchoolTrainings.unit.spec.jsx
@@ -1,0 +1,48 @@
+import React from 'react';
+import { findDOMNode } from 'react-dom';
+import { expect } from 'chai';
+import sinon from 'sinon';
+import ReactTestUtils from 'react-addons-test-utils';
+
+import { DefinitionTester } from '../../util/schemaform-utils.jsx';
+import uiSchema from '../../../src/js/edu-benefits/definitions/postHighSchoolTrainings';
+import definitions from 'vets-json-schema/dist/definitions.json';
+
+const schema = definitions.postHighSchoolTrainings;
+
+describe('Edu postHighSchoolTrainings', () => {
+  it('should render', () => {
+    const form = ReactTestUtils.renderIntoDocument(
+      <DefinitionTester
+          schema={schema}
+          data={[]}
+          definitions={definitions}
+          uiSchema={uiSchema}/>
+    );
+    const fields = Array.from(findDOMNode(form).querySelectorAll('input, select'));
+
+    expect(fields.length).to.equal(13);
+  });
+
+  it('should add another', () => {
+    const onSubmit = sinon.spy();
+    const form = ReactTestUtils.renderIntoDocument(
+      <DefinitionTester
+          schema={schema}
+          onSubmit={onSubmit}
+          data={[]}
+          definitions={definitions}
+          uiSchema={uiSchema}/>
+    );
+    const formDOM = findDOMNode(form);
+    ReactTestUtils.Simulate.change(formDOM.querySelector('#root_0_name'), {
+      target: {
+        value: 'A college name'
+      }
+    });
+    ReactTestUtils.Simulate.click(formDOM.querySelector('.va-growable-add-btn'));
+
+    expect(formDOM.querySelector('.va-growable-background').textContent)
+      .to.contain('A college name');
+  });
+});
